requirements.txt
easyocr>=1.7.1
opencv-python>=4.8.0
numpy<2.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
scikit-learn>=1.0.0
PyMuPDF>=1.22.0